Credentials: Fix PID SD-JWT with age attributes
