+++ shared/src/commonMain/kotlin/data/credentials/EuPidCredentialAttributeCategorization.kt
@@ -2,7 +2,6 @@ package data.credentials
 
 import at.asitplus.jsonpath.core.NormalizedJsonPath
 import at.asitplus.wallet.eupid.EuPidScheme
-import at.asitplus.wallet.eupidsdjwt.EuPidSdJwtScheme
 import data.PersonalDataCategory
 
 @Suppress("DEPRECATION")
@@ -80,54 +79,3 @@ object EuPidCredentialAttributeCategorization : CredentialAttributeCategorizatio
     },
 )
 
-object EuPidSdJwtCredentialAttributeCategorization : CredentialAttributeCategorization.Template(
-    mapOf(
-        PersonalDataCategory.IdentityData to with(EuPidSdJwtScheme.SdJwtAttributes) {
-            listOf(
-                GIVEN_NAME,
-                FAMILY_NAME,
-                BIRTH_DATE,
-                PORTRAIT,
-                NATIONALITIES,
-                SEX,
-            ).map { NormalizedJsonPath() + it to null }
-        },
-
-        PersonalDataCategory.BirthData to with(EuPidSdJwtScheme.SdJwtAttributes) {
-            listOf(
-                GIVEN_NAME_BIRTH,
-                FAMILY_NAME_BIRTH,
-                PLACE_OF_BIRTH_LOCALITY,
-                PLACE_OF_BIRTH_COUNTRY,
-                PLACE_OF_BIRTH_REGION,
-            ).map { NormalizedJsonPath() + it to null }
-        },
-
-        PersonalDataCategory.ResidenceData to with(EuPidSdJwtScheme.SdJwtAttributes) {
-            listOf(
-                ADDRESS_STREET,
-                ADDRESS_HOUSE_NUMBER,
-                ADDRESS_POSTAL_CODE,
-                ADDRESS_LOCALITY,
-                ADDRESS_COUNTRY,
-                ADDRESS_REGION,
-                ADDRESS_FORMATTED,
-            ).map { NormalizedJsonPath() + it to null }
-        },
-
-        PersonalDataCategory.Metadata to with(EuPidSdJwtScheme.SdJwtAttributes) {
-            listOf(
-                DOCUMENT_NUMBER,
-                ISSUANCE_DATE,
-                EXPIRY_DATE,
-                ISSUING_COUNTRY,
-                ISSUING_AUTHORITY,
-                ISSUING_JURISDICTION,
-                PERSONAL_ADMINISTRATIVE_NUMBER,
-            ).map { NormalizedJsonPath() + it to null }
-        },
-    ),
-    allAttributes = EuPidSdJwtScheme.claimNames.map {
-        NormalizedJsonPath() + it
-    },
-)
